Aula 06

+++ src/main/java/br/com/forum_hub/domain/resposta/RespostaService.java
@@ -2,8 +2,11 @@ package br.com.forum_hub.domain.resposta;
 
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
+import br.com.forum_hub.domain.autenticacao.HierarquiaService;
 import br.com.forum_hub.domain.topico.Status;
 import br.com.forum_hub.domain.topico.TopicoService;
 import br.com.forum_hub.domain.usuario.Usuario;
@@ -12,8 +15,12 @@ import jakarta.transaction.Transactional;
 
 @Service
 public class RespostaService {
+
     private final RespostaRepository repository;
     private final TopicoService topicoService;
+
+    @Autowired
+    HierarquiaService hierarquiaService;
 
     public RespostaService(RespostaRepository repository, TopicoService topicoService) {
         this.repository = repository;
@@ -24,11 +31,11 @@ public class RespostaService {
     public Resposta cadastrar(DadosCadastroResposta dados, Long idTopico, Usuario autor) {
         var topico = topicoService.buscarPeloId(idTopico);
 
-        if(!topico.estaAberto()) {
+        if (!topico.estaAberto()) {
             throw new RegraDeNegocioException("O tópico está fechado! Você não pode adicionar mais respostas.");
         }
 
-        if(topico.getQuantidadeRespostas() == 0) {
+        if (topico.getQuantidadeRespostas() == 0) {
             topico.alterarStatus(Status.RESPONDIDO);
         }
 
@@ -39,38 +46,49 @@ public class RespostaService {
     }
 
     @Transactional
-    public Resposta atualizar(DadosAtualizacaoResposta dados) {
+    public Resposta atualizar(DadosAtualizacaoResposta dados, Usuario logado) {
         var resposta = buscarPeloId(dados.id());
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, resposta.getAutor(), "ROLE_MODERADOR"))
+            throw new RegraDeNegocioException(
+                    "Você não pode editar essa resposta!");
         return resposta.atualizarInformacoes(dados);
     }
 
-    public List<Resposta> buscarRespostasTopico(Long id){
+    public List<Resposta> buscarRespostasTopico(Long id) {
         return repository.findByTopicoId(id);
     }
 
     @Transactional
-    public Resposta marcarComoSolucao(Long id) {
+    public Resposta marcarComoSolucao(Long id, Usuario logado) {
         var resposta = buscarPeloId(id);
 
         var topico = resposta.getTopico();
-        if(topico.getStatus() == Status.RESOLVIDO)
-            throw new RegraDeNegocioException("O tópico já foi solucionado! Você não pode marcar mais de uma resposta como solução.");
+
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_INSTRUTOR"))
+            throw new AccessDeniedException("Você não pode marcar essa resposta como solução!");
+
+        if (topico.getStatus() == Status.RESOLVIDO)
+            throw new RegraDeNegocioException(
+                    "O tópico já foi solucionado! Você não pode marcar mais de uma resposta como solução.");
 
         topico.alterarStatus(Status.RESOLVIDO);
         return resposta.marcarComoSolucao();
     }
 
     @Transactional
-    public void excluir(Long id) {
+    public void excluir(Long id, Usuario logado) {
         var resposta = buscarPeloId(id);
         var topico = resposta.getTopico();
 
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_MODERADOR"))
+            throw new RegraDeNegocioException(
+                    "Você não pode apagar essa resposta!");
         repository.deleteById(id);
 
         topico.decrementarRespostas();
         if (topico.getQuantidadeRespostas() == 0)
             topico.alterarStatus(Status.NAO_RESPONDIDO);
-        else if(resposta.ehSolucao())
+        else if (resposta.ehSolucao())
             topico.alterarStatus(Status.RESPONDIDO);
     }
